@@ -1,85 +1,79 @@
 {
   "name": "hope-just-found",
   "version": "0.0.0",
-  "description": "Description for HopeJustFound",
+  "description": "Description for hopeJustFound",
   "private": true,
   "license": "UNLICENSED",
   "cacheDirectories": [
     "node_modules"
   ],
   "dependencies": {
-    "@angular/common": "6.1.0",
-    "@angular/compiler": "6.1.0",
-    "@angular/core": "6.1.0",
-    "@angular/forms": "6.1.0",
-    "@angular/platform-browser": "6.1.0",
-    "@angular/platform-browser-dynamic": "6.1.0",
-    "@angular/router": "6.1.0",
-    "@fortawesome/angular-fontawesome": "0.2.0",
-    "@fortawesome/fontawesome-svg-core": "1.2.4",
-    "@fortawesome/free-solid-svg-icons": "5.3.1",
-    "@ng-bootstrap/ng-bootstrap": "3.0.0",
-    "angular-file-uploader": "^4.1.4",
+    "@angular/common": "7.0.0",
+    "@angular/compiler": "7.0.0",
+    "@angular/core": "7.0.0",
+    "@angular/forms": "7.0.0",
+    "@angular/platform-browser": "7.0.0",
+    "@angular/platform-browser-dynamic": "7.0.0",
+    "@angular/router": "7.0.0",
+    "@fortawesome/angular-fontawesome": "0.3.0",
+    "@fortawesome/fontawesome-svg-core": "1.2.8",
+    "@fortawesome/free-solid-svg-icons": "5.5.0",
+    "@ng-bootstrap/ng-bootstrap": "4.0.0",
     "bootstrap": "4.1.3",
     "core-js": "2.5.7",
     "moment": "2.22.2",
     "ng-jhipster": "0.5.4",
-    "ng2-file-upload": "^1.3.0",
     "ngx-cookie": "2.0.1",
     "ngx-infinite-scroll": "6.0.1",
     "ngx-webstorage": "2.0.1",
-    "ngxf-uploader": "^1.5.0",
     "reflect-metadata": "0.1.12",
-    "rxjs": "6.1.0",
+    "rxjs": "6.3.3",
     "swagger-ui": "2.2.10",
     "tslib": "1.9.3",
     "zone.js": "0.8.26"
   },
   "devDependencies": {
-    "@angular/cli": "6.1.2",
-    "@angular/compiler-cli": "6.1.0",
-    "@ngtools/webpack": "6.0.0",
-    "@types/jest": "22.2.3",
+    "@angular/cli": "7.0.2",
+    "@angular/compiler-cli": "7.0.0",
+    "@ngtools/webpack": "7.0.2",
+    "@types/jest": "23.3.5",
     "@types/node": "9.4.7",
     "angular-router-loader": "0.8.5",
     "angular2-template-loader": "0.6.2",
-    "browser-sync": "2.24.6",
+    "autoprefixer": "9.2.0",
+    "browser-sync": "2.26.3",
     "browser-sync-webpack-plugin": "2.2.2",
     "cache-loader": "1.2.2",
-    "codelyzer": "4.2.1",
+    "codelyzer": "4.5.0",
     "copy-webpack-plugin": "4.5.1",
     "css-loader": "0.28.10",
-    "exports-loader": "0.7.0",
     "file-loader": "1.1.11",
-    "fork-ts-checker-webpack-plugin": "0.4.1",
+    "fork-ts-checker-webpack-plugin": "0.4.10",
     "friendly-errors-webpack-plugin": "1.7.0",
-    "generator-jhipster": "5.5.0",
+    "generator-jhipster": "5.6.1",
     "html-loader": "0.5.5",
     "html-webpack-plugin": "3.2.0",
     "husky": "1.1.0",
-    "jest": "22.4.3",
+    "jest": "23.6.0",
     "jest-junit": "5.1.0",
-    "jest-preset-angular": "5.2.2",
+    "jest-preset-angular": "6.0.1",
     "jest-sonar-reporter": "2.0.0",
     "lint-staged": "7.3.0",
     "mini-css-extract-plugin": "0.4.2",
     "moment-locales-webpack-plugin": "1.0.5",
     "optimize-css-assets-webpack-plugin": "5.0.1",
-    "prettier": "1.11.1",
-    "proxy-middleware": "0.15.0",
-    "raw-loader": "0.5.1",
+    "prettier": "1.14.3",
     "rimraf": "2.6.1",
     "simple-progress-webpack-plugin": "1.1.2",
     "style-loader": "0.20.3",
-    "tapable": "1.0.0",
     "terser-webpack-plugin": "1.0.0",
     "thread-loader": "1.1.5",
     "to-string-loader": "1.1.5",
-    "ts-loader": "4.0.1",
-    "tslint": "5.9.1",
-    "tslint-config-prettier": "1.9.0",
+    "ts-loader": "4.5.0",
+    "tslint": "5.11.0",
+    "tslint-config-prettier": "1.15.0",
     "tslint-loader": "3.6.0",
-    "typescript": "2.7.2",
+    "typescript": "3.1.3",
     "postcss-loader": "2.1.1",
     "xml2js": "0.4.19",
     "generator-jhipster-docker": "2.5.0",
