@@ -18,11 +18,9 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.*;
-import org.springframework.web.multipart.MultipartFile;
 
 import javax.servlet.http.HttpServletRequest;
 import javax.validation.Valid;
-import java.io.File;
 import java.util.*;
 
 
@@ -64,16 +62,22 @@ public class AccountResource {
             throw new InvalidPasswordException();
         }
         User user = userService.registerUser(managedUserVM,
-                                             managedUserVM.getPassword(),
-                                             managedUserVM.getCompanyName(),
-                                             managedUserVM.getPhoneNumber(),
-                                             managedUserVM.getAddress(),
-                                             managedUserVM.getCity(),
-                                             managedUserVM.getState(),
-                                             managedUserVM.getZipCode(),
-                                             managedUserVM.getContactDays(),
-                                             managedUserVM.getContactTimes(),
-                                             managedUserVM.getAuthority());
+            managedUserVM.getPassword(),
+            managedUserVM.getCompanyName(),
+            managedUserVM.getCompanyWebsite(),
+            managedUserVM.getCompanyEIN(),
+            managedUserVM.getPhoneNumber(),
+            managedUserVM.getMailingAddress(),
+            managedUserVM.getMailingCity(),
+            managedUserVM.getMailingState(),
+            managedUserVM.getMailingZipCode(),
+            managedUserVM.getPhyisicalAddress(),
+            managedUserVM.getPhyisicalCity(),
+            managedUserVM.getPhyisicalState(),
+            managedUserVM.getPhyisicalZipCode(),
+            managedUserVM.getContactDays(),
+            managedUserVM.getContactTimes(),
+            managedUserVM.getAuthority());
 
         mailService.sendActivationEmail(user);
     }
