@@ -3,14 +3,14 @@ import { IUser } from 'app/core/user/user.model';
 
 export interface IDonation {
     id?: number;
-    type?: string;
+    goodsType?: string;
+    serviceType?: string;
+    imagesContentType?: string;
+    images?: any;
     initialDate?: Moment;
     expireDate?: Moment;
     condition?: string;
     description?: string;
-    experience?: string;
-    climate?: string;
-    intensity?: string;
     numberOfVolunteers?: number;
     user?: IUser;
 }
@@ -18,14 +18,14 @@ export interface IDonation {
 export class Donation implements IDonation {
     constructor(
         public id?: number,
-        public type?: string,
+        public goodsType?: string,
+        public serviceType?: string,
+        public imagesContentType?: string,
+        public images?: any,
         public initialDate?: Moment,
         public expireDate?: Moment,
         public condition?: string,
         public description?: string,
-        public experience?: string,
-        public climate?: string,
-        public intensity?: string,
         public numberOfVolunteers?: number,
         public user?: IUser
     ) {}
