@@ -14,22 +14,20 @@ describe('Component Tests', () => {
             data: of({ user: new User(1, 'user', 'first', 'last', 'first@last.com', true, 'en', ['ROLE_USER'], 'admin', null, null, null) })
         } as any) as ActivatedRoute;
 
-        beforeEach(
-            async(() => {
-                TestBed.configureTestingModule({
-                    imports: [HopeJustFoundTestModule],
-                    declarations: [UserMgmtDetailComponent],
-                    providers: [
-                        {
-                            provide: ActivatedRoute,
-                            useValue: route
-                        }
-                    ]
-                })
-                    .overrideTemplate(UserMgmtDetailComponent, '')
-                    .compileComponents();
+        beforeEach(async(() => {
+            TestBed.configureTestingModule({
+                imports: [HopeJustFoundTestModule],
+                declarations: [UserMgmtDetailComponent],
+                providers: [
+                    {
+                        provide: ActivatedRoute,
+                        useValue: route
+                    }
+                ]
             })
-        );
+                .overrideTemplate(UserMgmtDetailComponent, '')
+                .compileComponents();
+        }));
 
         beforeEach(() => {
             fixture = TestBed.createComponent(UserMgmtDetailComponent);
