@@ -16,22 +16,20 @@ describe('Component Tests', () => {
             data: of({ user: new User(1, 'user', 'first', 'last', 'first@last.com', true, 'en', ['ROLE_USER'], 'admin', null, null, null) })
         } as any) as ActivatedRoute;
 
-        beforeEach(
-            async(() => {
-                TestBed.configureTestingModule({
-                    imports: [HopeJustFoundTestModule],
-                    declarations: [UserMgmtUpdateComponent],
-                    providers: [
-                        {
-                            provide: ActivatedRoute,
-                            useValue: route
-                        }
-                    ]
-                })
-                    .overrideTemplate(UserMgmtUpdateComponent, '')
-                    .compileComponents();
+        beforeEach(async(() => {
+            TestBed.configureTestingModule({
+                imports: [HopeJustFoundTestModule],
+                declarations: [UserMgmtUpdateComponent],
+                providers: [
+                    {
+                        provide: ActivatedRoute,
+                        useValue: route
+                    }
+                ]
             })
-        );
+                .overrideTemplate(UserMgmtUpdateComponent, '')
+                .compileComponents();
+        }));
 
         beforeEach(() => {
             fixture = TestBed.createComponent(UserMgmtUpdateComponent);
@@ -40,71 +38,62 @@ describe('Component Tests', () => {
         });
 
         describe('OnInit', () => {
-            it(
-                'Should load authorities and language on init',
-                inject(
-                    [],
-                    fakeAsync(() => {
-                        // GIVEN
-                        spyOn(service, 'authorities').and.returnValue(of(['USER']));
+            it('Should load authorities and language on init', inject(
+                [],
+                fakeAsync(() => {
+                    // GIVEN
+                    spyOn(service, 'authorities').and.returnValue(of(['USER']));
 
-                        // WHEN
-                        comp.ngOnInit();
+                    // WHEN
+                    comp.ngOnInit();
 
-                        // THEN
-                        expect(service.authorities).toHaveBeenCalled();
-                        expect(comp.authorities).toEqual(['USER']);
-                    })
-                )
-            );
+                    // THEN
+                    expect(service.authorities).toHaveBeenCalled();
+                    expect(comp.authorities).toEqual(['USER']);
+                })
+            ));
         });
 
         describe('save', () => {
-            it(
-                'Should call update service on save for existing user',
-                inject(
-                    [],
-                    fakeAsync(() => {
-                        // GIVEN
-                        const entity = new User(123);
-                        spyOn(service, 'update').and.returnValue(
-                            of(
-                                new HttpResponse({
-                                    body: entity
-                                })
-                            )
-                        );
-                        comp.user = entity;
-                        // WHEN
-                        comp.save();
-                        tick(); // simulate async
+            it('Should call update service on save for existing user', inject(
+                [],
+                fakeAsync(() => {
+                    // GIVEN
+                    const entity = new User(123);
+                    spyOn(service, 'update').and.returnValue(
+                        of(
+                            new HttpResponse({
+                                body: entity
+                            })
+                        )
+                    );
+                    comp.user = entity;
+                    // WHEN
+                    comp.save();
+                    tick(); // simulate async
 
-                        // THEN
-                        expect(service.update).toHaveBeenCalledWith(entity);
-                        expect(comp.isSaving).toEqual(false);
-                    })
-                )
-            );
+                    // THEN
+                    expect(service.update).toHaveBeenCalledWith(entity);
+                    expect(comp.isSaving).toEqual(false);
+                })
+            ));
 
-            it(
-                'Should call create service on save for new user',
-                inject(
-                    [],
-                    fakeAsync(() => {
-                        // GIVEN
-                        const entity = new User();
-                        spyOn(service, 'create').and.returnValue(of(new HttpResponse({ body: entity })));
-                        comp.user = entity;
-                        // WHEN
-                        comp.save();
-                        tick(); // simulate async
+            it('Should call create service on save for new user', inject(
+                [],
+                fakeAsync(() => {
+                    // GIVEN
+                    const entity = new User();
+                    spyOn(service, 'create').and.returnValue(of(new HttpResponse({ body: entity })));
+                    comp.user = entity;
+                    // WHEN
+                    comp.save();
+                    tick(); // simulate async
 
-                        // THEN
-                        expect(service.create).toHaveBeenCalledWith(entity);
-                        expect(comp.isSaving).toEqual(false);
-                    })
-                )
-            );
+                    // THEN
+                    expect(service.create).toHaveBeenCalledWith(entity);
+                    expect(comp.isSaving).toEqual(false);
+                })
+            ));
         });
     });
 });
